requirements.txt
alembic==1.8.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
